Versioning

+++ PhotonBox/src/Math/Math.h
@@ -25,7 +25,7 @@ namespace Math
 		// roll (x-axis rotation)
 		double sinr = +2.0 * (q.w * q.x + q.y * q.z);
 		double cosr = +1.0 - 2.0 * (q.x * q.x + q.y * q.y);
-		roll = atan2(sinr, cosr);
+		roll = (float)atan2(sinr, cosr);
 
 		// pitch (y-axis rotation)
 		/*
@@ -36,15 +36,15 @@ namespace Math
 		
 		double sinp = +2.0 * (q.w * q.y - q.z * q.x);
 		if (fabs(sinp) >= 1)
-			pitch = copysign(PI / 2, sinp); // use 90 degrees if out of range
+			pitch = (float)copysign(PI / 2, sinp); // use 90 degrees if out of range
 		else
-			pitch = asin(sinp);
+			pitch = (float)asin(sinp);
 		
 
 		// yaw (z-axis rotation)
 		double siny = +2.0 * (q.w * q.z + q.x * q.y);
 		double cosy = +1.0 - 2.0 * (q.y * q.y + q.z * q.z);
-		yaw = atan2(siny, cosy);
+		yaw = (float)atan2(siny, cosy);
 
 		return Vector3f(roll, pitch, yaw);
 	}
@@ -57,12 +57,12 @@ namespace Math
 		// roll (x-axis rotation)
 		double sinr = +2.0 * (q.getW() * q.getX() + q.getY() * q.getZ());
 		double cosr = +1.0 - 2.0 * (q.getX() * q.getX() + q.getY() * q.getY());
-		roll = atan2(sinr, cosr);
+		roll = (float)atan2(sinr, cosr);
 
 		// pitch (y-axis rotation)
 		float m = (std::max)(-2.0f * (q.getX() * q.getZ() - q.getW() * q.getY()), -1.0f);
 		float t = (std::min)(m, 1.0f);
-		pitch = asin(t);
+		pitch = (float)asin(t);
 
 		/*
 		double sinp = +2.0 * (q.getW() * q.getY() - q.getZ() * q.getX());
@@ -75,7 +75,7 @@ namespace Math
 		// yaw (z-axis rotation)
 		double siny = +2.0 * (q.getW()* q.getZ() + q.getX()* q.getY());
 		double cosy = +1.0 - 2.0 * (q.getY()* q.getY() + q.getZ()* q.getZ());
-		yaw = atan2(siny, cosy);
+		yaw = (float)atan2(siny, cosy);
 
 		return Vector3f(roll, pitch, yaw);
 	}
